--- data/Duke.txt
@@ -1,4 +1,0 @@
-Do these soon!!
-[T][X]  hw
-[T][ ]  deadline
-[D][X]  hw (by Oct 15 2019)
