+++ src/main/java/task/Deadline.java
@@ -1,10 +1,16 @@
-public class Event extends Task{
+package task;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+public class Deadline extends Task{
     public String type;
-    public String date_time = "";
-    public String task;
+    public String task = "";
+    public LocalDate date_time;
+    public String date_time2;
 
 
-    public Event (String message) {
+    public Deadline (String message) {
         super(message);
         this.set_date_time();
         this.set_task();
@@ -12,8 +18,7 @@ public class Event extends Task{
     }
 
 
-    //constructor for Duke.txt
-    public Event (String message, boolean b) {
+    public Deadline (String message, boolean b) {
         super(message);
         this.set_date_time2();
         this.set_task2();
@@ -26,12 +31,17 @@ public class Event extends Task{
     }
 
     @Override
+    public void set_type() {
+        this.type = "D";
+    }
+
+    @Override
     public void set_task() {
         int start_index = 0;
         int end_index = 0;
         for (int i = 0; i < this.message.length(); i++) {
-            if (this.message.substring(i, i+1).equals("t")) {
-                start_index = i + 2;
+            if (this.message.substring(i, i+1).equals("d")) {
+                start_index = i + 8;
                 break;
             }
         }
@@ -41,17 +51,16 @@ public class Event extends Task{
                 break;
             }
         }
-        this.task = " " + message.substring(start_index,end_index) + this.getDate_time();
+        this.task = message.substring(start_index,end_index) + " (by " + this.get_date_time() + ")";
     }
 
-    //set task for Duke.txt
     @Override
     public void set_task2() {
         int start_index = 0;
         int end_index = 0;
         for (int i = 0; i < this.message.length(); i++) {
-            if (this.message.substring(i, i+1).equals("e")) {
-                start_index = i + 6;
+            if (this.message.substring(i, i+1).equals("d")) {
+                start_index = i + 9;
                 break;
             }
         }
@@ -61,18 +70,12 @@ public class Event extends Task{
                 break;
             }
         }
-        this.task = message.substring(start_index,end_index) + this.getDate_time();
+        this.task = message.substring(start_index,end_index) + " " + this.get_date_time2();
     }
-
 
     @Override
     public String get_task() {
         return this.task;
-    }
-
-    @Override
-    public void set_type() {
-        this.type = "E";
     }
 
     @Override
@@ -84,22 +87,21 @@ public class Event extends Task{
                 break;
             }
         }
-        this.date_time = " (at " + message.substring(start_index,message.length()) + ")";
+        String date_timing = message.substring(start_index);
+        this.date_time = LocalDate.parse(date_timing);
     }
 
     public void set_date_time2() {
-        int start_index = 0;
-        for (int i = 0; i < this.message.length(); i++) {
-            if (this.message.substring(i, i+1).equals("(")) {
-                start_index = i - 1;
-                break;
-            }
-        }
-        this.date_time = message.substring(start_index,message.length());
+        int length = this.message.length();
+        this.date_time2 = "(" + message.substring(length - 15);
     }
 
-    @Override
-    public String getDate_time() {
-        return date_time;
+    public String get_date_time() {
+        return this.date_time.format(DateTimeFormatter.ofPattern("MMM d yyyy"));
     }
+
+    public String get_date_time2() {
+        return this.date_time2;
+    }
+
 }
